feat: getting started revamp
