+++ packages/webapp/package.json
@@ -63,7 +63,6 @@
         "@tailwindcss/forms": "^0.5.3",
         "@types/jest": "^27.5.2",
         "@types/js-cookie": "^3.0.3",
-        "@types/node": "18.7.6",
         "@types/react": "18.2.58",
         "@types/react-dom": "18.2.19",
         "@types/react-helmet": "^6.1.8",
